+++ jina/executors/indexers/query/__init__.py
@@ -1,4 +1,4 @@
-from typing import Iterable
+from typing import Iterable, Optional, Dict
 
 from jina.executors.indexers import BaseIndexer
 
@@ -6,7 +6,22 @@ from jina.executors.indexers import BaseIndexer
 class BaseQueryIndexer(BaseIndexer):
     """An indexer only for querying. It only reads once (at creation time, from a dump)"""
 
-    def load_dump(self, dump_path):
+    def _post_init_wrapper(
+        self,
+        _metas: Optional[Dict] = None,
+        _requests: Optional[Dict] = None,
+        fill_in_metas: bool = True,
+    ) -> None:
+        super()._post_init_wrapper(_metas, _requests, fill_in_metas)
+        self.dump_path = _metas.get('dump_path')
+        # TODO fix this
+        # don't do this for Compounds, as the _components
+        # are not yet set at this stage.
+        # for Compound we use a `_post_components`
+        if self.dump_path and not hasattr(self, 'components'):
+            self._load_dump(self.dump_path)
+
+    def _load_dump(self, dump_path):
         """Load the dump at the dump_path
 
         :param dump_path: the path of the dump"""
