+++ jina/executors/indexers/query/compound.py
@@ -1,5 +1,3 @@
-from typing import Optional, Dict
-
 from jina.executors.compound import CompoundExecutor
 from jina.executors.indexers.query import BaseQueryIndexer
 
@@ -10,19 +8,14 @@ class QueryCompoundExecutor(CompoundExecutor, BaseQueryIndexer):
     :param dump_path: the path to initialize from
     """
 
-    def _post_init_wrapper(
-        self,
-        _metas: Optional[Dict] = None,
-        _requests: Optional[Dict] = None,
-        fill_in_metas: bool = True,
-    ) -> None:
-        super()._post_init_wrapper(_metas, _requests, fill_in_metas)
-        self.dump_path = _metas.get('dump_path')
-
+    # TODO fix this
+    # don't do this for Compounds, as the _components
+    # are not yet set at this stage.
+    # for Compound we use a `_post_components`
     def _post_components(self):
-        self.load_dump(self.dump_path)
+        self._load_dump(self.dump_path)
 
-    def load_dump(self, dump_path, *args, **kwargs):
+    def _load_dump(self, dump_path, *args, **kwargs):
         """Loads the data in the indexer
 
         :param dump_path: the path to the dump
@@ -30,4 +23,4 @@ class QueryCompoundExecutor(CompoundExecutor, BaseQueryIndexer):
         :param kwargs: passed to the inner Indexer's load_dump
         """
         for c in self.components:
-            c.load_dump(dump_path)
+            c._load_dump(dump_path)
